# PetStore/static/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "This is a sample server Petstore server. Added a Bidding Option to it",
    "version": "1.0.5",
    "title": "Swagger Petstore With Bids",
    "termsOfService": "http://swagger.io/terms/",
    "contact": {
      "email": "apiteam@swagger.io"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
...
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/Bids/All": {
      "get": {
        "tags": [
          "Bids"
        ],
        "summary": "Return All Bids",
        "description": "",
        "produces": [
          "application/json",
          "application/xml"
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": " "
            }
          },
          "404": {
            "description": "No Bids Found"
          }
        }
      }
    },
    "/Bids": {
      "get": {
        "tags": [
          "Bids"
        ],
        "summary": "Return all Bids on specific pet",
        "description": "",
        "produces": [
          "application/json",
          "application/xml"
        ],
        "parameters": [
          {
            "name": "pet_id",
            "in": "query",
            "description": "The ID that needs to be fetched. Use 1 for testing. ",
            "required": true,
            "type": "string",
            "default": "1"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
            }
          },
          "404": {
            "description": "No Bids Found"
          },
          "400": {
            "description": "Bad Data Posted"
          }
        }
      }
    },
    "/Bids/Insert": {
      "post": {
        "tags": [
          "Bids"
        ],
        "summary": "Bid on a specific pet from a specific user",
        "parameters": [
          {
            "name": "info",
            "in": "body",
            "required": true,
            "schema": {
                  "$ref": "#/definitions/Bids"

              }
            }

        ],
        "description": "Users can bid on a specific pet, a specific amount of money.For Testing purposes use  pet_id [1,2,3],user_id[4,5]",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Bids"
            }
          },
          "400": {
            "description": "Failed. Bad post data."
          }
        }
      }
    }
  },
  "securityDefinitions": {
    "api_key": {
      "type": "apiKey",
      "name": "api_key",
      "in": "header"
    },
    "petstore_auth": {
      "type": "oauth2",
      "authorizationUrl": "https://petstore.swagger.io/oauth/authorize",
      "flow": "implicit",
      "scopes": {
        "read:pets": "read your pets",
        "write:pets": "modify pets in your account"
      }
    }
  },
  "definitions": {
...
    "Bids": {
      "type": "object",
      "properties": {
        "pet_id": {
          "$ref": "#/definitions/Pet/properties/id"
        },
        "user_id": {
          "$ref": "#/definitions/User/properties/id"
        },
        "money": {
          "type": "integer",
          "format": "number"
        }

      },
      "xml": {
        "name": "Bids"
      }
    },
...
  },
...
}